# Conflicts:
#	src/broker/module/client/broker_client_nodes.cc

+++ src/sdk/module/default/simple_storage.cc
@@ -39,13 +39,15 @@ void SimpleStorage::destroy_impl() {
 
 void SimpleStorage::clear() {
   fs::path p(storage_default);
-  for (auto&& x : fs::directory_iterator(p)) {
-    std::wstring path_entry(x.path().stem().wstring());
-    SimpleStorage simple_storage;
-    shared_ptr_<StorageBucket> storage_bucket;
-    storage_bucket = simple_storage.get_shared_bucket(url_decode(
-        std::wstring_convert<std::codecvt_utf8<wchar_t>>{}.to_bytes(path_entry)));
-    storage_bucket->remove_all();
+  if(fs::exists(p)) {
+    for (auto &&x : fs::directory_iterator(p)) {
+      std::wstring path_entry(x.path().stem().wstring());
+      SimpleStorage simple_storage;
+      shared_ptr_<StorageBucket> storage_bucket;
+      storage_bucket = simple_storage.get_shared_bucket(url_decode(
+          std::wstring_convert<std::codecvt_utf8<wchar_t>>{}.to_bytes(path_entry)));
+      storage_bucket->remove_all();
+    }
   }
 }
